refactor: core/residual → simulation/_residual、core/jacobian → verification/jacobian に移動
core/ をユーザー API 専用に純化し、依存方向を verification → simulation → core に統一。
JacobianBlocks / ad_jacobian_blocks は manforge.verification から公開。
