+++ src/manforge/core/__init__.py
@@ -7,7 +7,6 @@ from manforge.core.dimension import (
     UNIAXIAL_1D,
 )
 from manforge.core.result import ReturnMappingResult, StressUpdateResult
-from manforge.core.jacobian import JacobianBlocks, ad_jacobian_blocks
 
 __all__ = [
     "Implicit",
@@ -24,6 +23,4 @@ __all__ = [
     "UNIAXIAL_1D",
     "ReturnMappingResult",
     "StressUpdateResult",
-    "JacobianBlocks",
-    "ad_jacobian_blocks",
 ]
