+++ src/manforge/simulation/_residual.py
